parser/tests/tree_construction/basic.dat

#data
</html><frameset></frameset>
#errors
(1,7): expected-doctype-but-got-end-tag
(1,17): expected-eof-but-got-start-tag
(1,17): unexpected-start-tag
#document
| <html>
|   <head>
|   <frameset>

#data
</body><frameset></frameset>
#errors
(1,7): expected-doctype-but-got-end-tag
(1,17): unexpected-start-tag-after-body
(1,17): unexpected-start-tag
#document
| <html>
|   <head>
|   <frameset>
